Refactor and cleanup pytest suite
- Consolidated Frequency Counter widget tests into `tests/logic_verification/test_frequency_counter_widget.py`.
- Renamed `test_frequency_counter.py` to `test_analysis_frequency.py` to clarify it tests AudioCalc logic.
- Moved `test_lockin_performance.py` to `tests/benchmarks/benchmark_lockin_performance.py`.
- Deleted redundant/empty tests: `test_network_analyzer.py`, `test_lockin_pid.py`, `test_frequency_counter_*.py` (original files).
- Verified all tests pass.
